Consolidate interface
