+++ build.gradle
@@ -162,7 +162,7 @@ configurations {
 dependencies {
     implementation "org.opensearch.plugin:geo:${opensearch_version}"
     api project(":libs:h3")
-    api project(":geospatial-common")
+    api project(":geospatial-client")
     yamlRestTestRuntimeOnly "org.apache.logging.log4j:log4j-core:${versions.log4j}"
     testImplementation "org.hamcrest:hamcrest:${versions.hamcrest}"
     testImplementation 'org.json:json:20231013'
