+++ client/src/main/java/org/opensearch/geospatial/action/IpEnrichmentResponse.java
@@ -15,7 +15,6 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.UncheckedIOException;
-import java.util.UUID;
 
 
 public class IpEnrichmentResponse extends ActionResponse {
